refactor: 예약 서비스 내에서 예약 생성하도록 수정
- 내부에서 총 가격 합계 체크하도록 수정
