+++ src/main/java/com/culture/ticketing/booking/api/BookingController.java
@@ -1,6 +1,6 @@
 package com.culture.ticketing.booking.api;
 
-import com.culture.ticketing.booking.application.BookingFacadeService;
+import com.culture.ticketing.booking.application.BookingService;
 import com.culture.ticketing.booking.application.dto.BookingSaveRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,15 +11,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/bookings")
 public class BookingController {
 
-    private final BookingFacadeService bookingFacadeService;
+    private final BookingService bookingService;
 
-    public BookingController(BookingFacadeService bookingFacadeService) {
-        this.bookingFacadeService = bookingFacadeService;
+    public BookingController(BookingService bookingService) {
+        this.bookingService = bookingService;
     }
 
     @PostMapping
     public void createBooking(@RequestBody BookingSaveRequest request) {
 
-        bookingFacadeService.createBooking(request);
+        bookingService.createBooking(request);
     }
 }
